feat(divider): isolate allowed color tokens & update imports
