+++ src/components/divider/divider.tsx
@@ -1,44 +1,41 @@
 import React from 'react';
-import { View, StyleProp, ViewStyle } from 'react-native';
+import { View } from 'react-native';
 
+import { useThemeColor } from '../../utils/use-theme-color.hook';
+
+import { ALLOWED_DIVIDER_COLORS, AllowedColor, AllowedShade } from './divider.colors';
 import { DividerOrientation, DividerDashStyle, useDividerStyles } from './divider.styles';
-
-import { ALLOWED_DIVIDER_COLORS } from '@/app-theme';
-import { useThemeColor } from '@/utils/use-theme-color.hook';
-
-type AllowedColor = keyof typeof ALLOWED_DIVIDER_COLORS;
-type AllowedShade = (typeof ALLOWED_DIVIDER_COLORS)[AllowedColor];
 
 export interface DividerProps {
   orientation?: DividerOrientation;
   thickness?: number;
-  style?: StyleProp<ViewStyle>;
   length?: number | string;
   dashStyle?: DividerDashStyle;
+  testID?: string;
 }
 
-const Divider: React.FC<DividerProps> = ({
-  orientation = DividerOrientation.Horizontal,
-  thickness = 1,
-  style,
-  length,
-  dashStyle = DividerDashStyle.Solid,
-}) => {
+const Divider: React.FC<DividerProps> = ({ orientation, thickness, length, dashStyle, testID }) => {
   const color: AllowedColor = 'primary';
   const shade: AllowedShade = ALLOWED_DIVIDER_COLORS[color];
-
   const dividerColor = useThemeColor(`${color}.${shade}`);
-  const styles = useDividerStyles({ orientation, thickness, length, dashStyle, dividerColor });
 
-  return <View testID="divider" style={[styles.divider, style]} />;
+  const styles = useDividerStyles({
+    orientation: orientation!,
+    thickness: thickness!,
+    length,
+    dashStyle: dashStyle!,
+    dividerColor,
+  });
+
+  return <View testID={testID ?? 'divider'} style={[styles.divider]} />;
 };
 
 Divider.defaultProps = {
   orientation: DividerOrientation.Horizontal,
   thickness: 1,
-  style: undefined,
   length: undefined,
   dashStyle: DividerDashStyle.Solid,
+  testID: 'divider',
 };
 
 export default Divider;
